refactor: move deriveDisplayUrls from types/ to utils/ per CONVENTIONS.md
types/ is for cross-domain shared types only, not runtime helpers.
deriveDisplayUrls is a pure function — belongs in utils/.
